Fix a narrowing conversion warning.

+++ firmware/application/ui_setup.cpp
@@ -121,7 +121,7 @@ SetFrequencyCorrectionView::SetFrequencyCorrectionView(
 	} });
 
 	SetFrequencyCorrectionModel model {
-		portapack::persistent_memory::correction_ppb() / 1000
+		(int8_t) (portapack::persistent_memory::correction_ppb() / 1000)
 	};
 
 	form_init(model);
